Add commands mod
