+++ src/commands/mod.rs
@@ -1,0 +1,5 @@
+mod build;
+mod run;
+
+pub use build::*;
+pub use run::*;
